+++ src/common/mapper.ts
@@ -2002,9 +2002,10 @@ export class Mapper extends EventEmitter {
                 const rl = rows.length;
                 this.emit('import-progress', 0);
                 for (let r = 0; r < rl; r++) {
+                    rows[r].ID = parseInt(rows[r].ID, 10);
                     if (rooms[rows[r].ID]) {
                         rooms[rows[r].ID].exits[rows[r].Exit] = {
-                            num: rows[r].DestID,
+                            num: parseInt(rows[r].DestID, 10),
                             isdoor: rows[r].IsDoor,
                             isclosed: rows[r].IsClosed
                         };
@@ -2022,7 +2023,7 @@ export class Mapper extends EventEmitter {
                         }
                         rooms[rows[r].ID].exits = {};
                         rooms[rows[r].ID].exits[rows[r].Exit] = {
-                            num: rows[r].DestID,
+                            num: parseInt(rows[r].DestID, 10),
                             isdoor: rows[r].IsDoor,
                             isclosed: rows[r].IsClosed
                         };
@@ -2041,9 +2042,10 @@ export class Mapper extends EventEmitter {
                 const rl = rows.length;
                 this.emit('import-progress', 0);
                 for (let r = 0; r < rl; r++) {
+                    rows[r].ID = parseInt(rows[r].ID, 10);
                     if (rooms[rows[r].ID]) {
                         rooms[rows[r].ID].exits[rows[r].Exit] = {
-                            num: rows[r].DestID,
+                            num: parseInt(rows[r].DestID, 10),
                             isdoor: rows[r].IsDoor,
                             isclosed: rows[r].IsClosed
                         };
@@ -2061,7 +2063,7 @@ export class Mapper extends EventEmitter {
                         }
                         rooms[rows[r].ID].exits = {};
                         rooms[rows[r].ID].exits[rows[r].Exit] = {
-                            num: rows[r].DestID,
+                            num: parseInt(rows[r].DestID, 10),
                             isdoor: rows[r].IsDoor,
                             isclosed: rows[r].IsClosed
                         };